Project Starlink/requirements.txt
Flask>=2.0
python-dotenv>=0.15
requests>=2.20
Werkzeug>=2.2.2 